part 1 - consistently handle dose-respose plot layout and dataset
